requirements.txt
python-dotenv~=1.2.1
gymnasium~=1.1.1
pygame~=2.6.1
opencv-python~=4.12.0.88
numpy~=2.0.2
matplotlib~=3.9.4
torch~=2.8.0
openai~=2.6.1
